Refactor embeddings, change way periodicity is specified
